+++ app/src/main/java/com/qdhc/ny/entity/Project.java
@@ -27,7 +27,7 @@ public class Project implements Serializable {
     /**
      * 项目进度
      */
-    private int process;
+    private float process;
 
     /**
      * 项目创建者Id
@@ -93,11 +93,11 @@ public class Project implements Serializable {
         this.info = info;
     }
 
-    public int getProcess() {
+    public float getProcess() {
         return process;
     }
 
-    public void setProcess(int process) {
+    public void setProcess(float process) {
         this.process = process;
     }
